+++ packages/broker/src/plugins/mqtt/MqttServer.ts
@@ -1,9 +1,10 @@
 import * as aedes from 'aedes'
 import * as net from 'net'
 import util from 'util'
-import { ISubscription, IPublishPacket } from 'mqtt-packet'
+import { ISubscription } from 'mqtt-packet'
 import { Logger } from '@streamr/utils'
 import { ApiAuthentication, isValidAuthentication } from '../../apiAuthentication'
+import Aedes from 'aedes'
 
 const logger = new Logger(module)
 
@@ -20,16 +21,16 @@ export class MqttServer {
     private static NOT_AUTHORIZED = 5
 
     private readonly port: number
-    private readonly aedes: aedes.Aedes
+    private readonly aedes: Aedes
     private server?: net.Server
     private listener?: MqttServerListener
 
     constructor(port: number, apiAuthentication?: ApiAuthentication) {
         this.port = port
-        this.aedes = aedes.Server({
+        this.aedes = new Aedes({
             authenticate: MqttServer.createAuthenicationHandler(apiAuthentication)
         })
-        this.aedes.on('publish', (packet: IPublishPacket, client: aedes.Client) => {
+        this.aedes.on('publish', (packet, client) => {
             if (client !== null) {  // is null if the this server sent the message
                 this.listener?.onMessageReceived(packet.topic, packet.payload.toString(), client.id)
             }
@@ -50,7 +51,7 @@ export class MqttServer {
     async start(): Promise<void> {
         this.server = net.createServer(this.aedes.handle)
         await util.promisify((callback: any) => this.server!.listen(this.port, callback))()
-        logger.info(`MQTT server listening on port ${this.port}`)
+        logger.info(`Started MQTT server on port ${this.port}`)
     }
 
     async stop(): Promise<void> {
@@ -58,7 +59,7 @@ export class MqttServer {
             const closeAedes = util.promisify((callback: any) => this.aedes.close(callback))()
             const closeServer = util.promisify((callback: any) => this.server!.close(callback))()
             await Promise.all([closeAedes, closeServer])
-            logger.info('MQTT server stopped')
+            logger.info('Stopped MQTT server')
         }
     }
 
@@ -73,7 +74,7 @@ export class MqttServer {
         }
         this.aedes.publish(packet, (error?: Error) => {
             if (error) {
-                logger.warn(`Publish error: ${error}`)
+                logger.warn('Failed to publish', { error, topic })
             }
         })
     }