+++ packages/client/src/Config.ts
@@ -299,6 +299,7 @@ export interface StreamrClientConfig {
         /** Some TheGraph instance, that indexes the streamr registries */
         theGraphUrl?: string
         maxConcurrentCalls?: number
+        pollInterval?: number
     }
 
     /**