+++ packages/client/src/registry/StreamStorageRegistry.ts
@@ -7,9 +7,9 @@ import { Stream } from '../Stream'
 import { getStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
 import { StreamID, toStreamID } from '@streamr/protocol'
 import { StreamIDBuilder } from '../StreamIDBuilder'
-import { waitForTx, queryAllReadonlyContracts } from '../utils/contract'
+import { waitForTx, queryAllReadonlyContracts, initContractEventGateway } from '../utils/contract'
 import { SynchronizedGraphQLClient } from '../utils/SynchronizedGraphQLClient'
-import { StreamrClientEventEmitter, StreamrClientEvents, initEventGateway } from '../events'
+import { StreamrClientEventEmitter } from '../events'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
 import { ContractFactory } from '../ContractFactory'
 import { EthereumAddress, Logger, toEthereumAddress } from '@streamr/utils'
@@ -71,35 +71,32 @@ export class StreamStorageRegistry {
                 'streamStorageRegistry'
             ) as StreamStorageRegistryContract
         })
-        this.initStreamAssignmentEventListener('addToStorageNode', 'Added', eventEmitter)
-        this.initStreamAssignmentEventListener('removeFromStorageNode', 'Removed', eventEmitter)
+        this.initStreamAssignmentEventListeners(eventEmitter, loggerFactory)
     }
 
-    private initStreamAssignmentEventListener(
-        clientEvent: keyof StreamrClientEvents,
-        contractEvent: string,
-        eventEmitter: StreamrClientEventEmitter
-    ) {
+    private initStreamAssignmentEventListeners(eventEmitter: StreamrClientEventEmitter, loggerFactory: LoggerFactory) {
         const primaryReadonlyContract = this.streamStorageRegistryContractsReadonly[0]
-        type Listener = (streamId: string, nodeAddress: string, extra: any) => void
-        initEventGateway(
-            clientEvent,
-            (emit: (payload: StorageNodeAssignmentEvent) => void) => {
-                const listener = (streamId: string, nodeAddress: string, extra: any) => {
-                    emit({
-                        streamId: toStreamID(streamId),
-                        nodeAddress: toEthereumAddress(nodeAddress),
-                        blockNumber: extra.blockNumber
-                    })
-                }
-                primaryReadonlyContract.on(contractEvent, listener)
-                return listener
-            },
-            (listener: Listener) => {
-                primaryReadonlyContract.off(contractEvent, listener)
-            },
-            eventEmitter
-        )
+        const transformation = (streamId: string, nodeAddress: string, extra: any) => ({
+            streamId: toStreamID(streamId),
+            nodeAddress: toEthereumAddress(nodeAddress),
+            blockNumber: extra.blockNumber
+        })
+        initContractEventGateway({
+            sourceName: 'Added', 
+            sourceEmitter: primaryReadonlyContract,
+            targetName: 'addToStorageNode',
+            targetEmitter: eventEmitter,
+            transformation,
+            loggerFactory
+        })
+        initContractEventGateway({
+            sourceName: 'Removed', 
+            sourceEmitter: primaryReadonlyContract,
+            targetName: 'removeFromStorageNode',
+            targetEmitter: eventEmitter,
+            transformation,
+            loggerFactory
+        })
     }
 
     private async connectToContract() {
@@ -116,7 +113,7 @@ export class StreamStorageRegistry {
 
     async addStreamToStorageNode(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<void> {
         const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
-        this.logger.debug('adding stream %s to node %s', streamId, nodeAddress)
+        this.logger.debug('Add stream to storage node', { streamId, nodeAddress })
         await this.connectToContract()
         const ethersOverrides = getStreamRegistryOverrides(this.config)
         await waitForTx(this.streamStorageRegistryContract!.addStorageNode(streamId, nodeAddress, ethersOverrides))
@@ -124,7 +121,7 @@ export class StreamStorageRegistry {
 
     async removeStreamFromStorageNode(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<void> {
         const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
-        this.logger.debug('removing stream %s from node %s', streamId, nodeAddress)
+        this.logger.debug('Remove stream from storage node', { streamId, nodeAddress })
         await this.connectToContract()
         const ethersOverrides = getStreamRegistryOverrides(this.config)
         await waitForTx(this.streamStorageRegistryContract!.removeStorageNode(streamId, nodeAddress, ethersOverrides))
@@ -132,14 +129,14 @@ export class StreamStorageRegistry {
 
     async isStoredStream(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<boolean> {
         const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
-        this.logger.debug('querying if stream %s is stored in storage node %s', streamId, nodeAddress)
+        this.logger.debug('Check if stream is stored in storage node', { streamId, nodeAddress })
         return queryAllReadonlyContracts((contract: StreamStorageRegistryContract) => {
             return contract.isStorageNodeOf(streamId, nodeAddress)
         }, this.streamStorageRegistryContractsReadonly)
     }
 
     async getStoredStreams(nodeAddress: EthereumAddress): Promise<{ streams: Stream[], blockNumber: number }> {
-        this.logger.debug('getting stored streams of node %s', nodeAddress)
+        this.logger.debug('Get stored streams of storage node', { nodeAddress })
         const blockNumbers: number[] = []
         const res = await collect(this.graphQLClient.fetchPaginatedResults(
             (lastId: string, pageSize: number) => {
@@ -181,7 +178,7 @@ export class StreamStorageRegistry {
         let queryResults: NodeQueryResult[]
         if (streamIdOrPath !== undefined) {
             const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
-            this.logger.debug('getting storage nodes of stream %s', streamId)
+            this.logger.debug('Get storage nodes of stream', { streamId })
             queryResults = await collect(this.graphQLClient.fetchPaginatedResults<NodeQueryResult>(
                 (lastId: string, pageSize: number) => {
                     const query = `{
@@ -202,7 +199,7 @@ export class StreamStorageRegistry {
                 }
             ))
         } else {
-            this.logger.debug('getting all storage nodes')
+            this.logger.debug('Get all storage nodes')
             queryResults = await collect(this.graphQLClient.fetchPaginatedResults<NodeQueryResult>(
                 (lastId: string, pageSize: number) => {
                     const query = `{