+++ packages/network/src/connection/ws/BrowserClientWsEndpoint.ts
@@ -3,6 +3,9 @@ import { PeerId, PeerInfo } from '../PeerInfo'
 import { DisconnectionCode, DisconnectionReason } from "./AbstractWsEndpoint"
 import { BrowserClientWsConnection, BrowserWebSocketConnectionFactory } from './BrowserClientWsConnection'
 import { AbstractClientWsEndpoint, HandshakeValues } from "./AbstractClientWsEndpoint"
+import { Logger } from '@streamr/utils'
+
+const logger = new Logger(module)
 
 export default class BrowserClientWsEndpoint extends AbstractClientWsEndpoint<BrowserClientWsConnection> {
     protected doConnect(serverUrl: string, serverPeerInfo: PeerInfo): Promise<PeerId> {
@@ -27,7 +30,7 @@ export default class BrowserClientWsEndpoint extends AbstractClientWsEndpoint<Br
                 }
 
             } catch (err) {
-                this.logger.trace('failed to connect to %s, error: %o', serverUrl, err)
+                logger.trace('Failed to connect to server', { serverUrl, err })
                 reject(err)
             }
         })
@@ -48,9 +51,9 @@ export default class BrowserClientWsEndpoint extends AbstractClientWsEndpoint<Br
         ws.onclose = (event) => {
             this.onClose(connection, event.code, event.reason as DisconnectionReason)
             if (event.code === DisconnectionCode.DUPLICATE_SOCKET) {
-                this.logger.warn('Connection refused: Duplicate nodeId detected, are you running multiple nodes with the same private key?')
+                logger.warn('Refused connection (Duplicate nodeId detected, are you running multiple nodes with the same private key?)')
             } else if (event.code === DisconnectionCode.INVALID_PROTOCOL_MESSAGE) {
-                this.logger.warn('Connection refused: Invalid protocol message format detected, are you running an outdated version?')
+                logger.warn('Refused connection (Invalid protocol message format detected, are you running an outdated version?)')
             }
         }