+++ packages/network/src/connection/ws/ServerWsEndpoint.ts
@@ -1,6 +1,6 @@
 import { PeerId, PeerInfo } from '../PeerInfo'
 import { AbstractWsEndpoint, DisconnectionCode, DisconnectionReason, } from "./AbstractWsEndpoint"
-import { staticLogger, ServerWsConnection } from './ServerWsConnection'
+import { ServerWsConnection } from './ServerWsConnection'
 import fs from 'fs'
 import net from 'net'
 import https from 'https'
@@ -9,6 +9,7 @@ import WebSocket from 'ws'
 import { once } from 'events'
 import { v4 } from 'uuid'
 import { Duplex } from "stream"
+import { Logger } from '@streamr/utils'
 
 interface HostPort {
     hostname: string
@@ -17,6 +18,8 @@ interface HostPort {
 type UnixSocket = string
 
 export type HttpServerConfig = HostPort | UnixSocket
+
+const logger = new Logger(module)
 
 export class ServerWsEndpoint extends AbstractWsEndpoint<ServerWsConnection> {
     private readonly serverUrl: string
@@ -46,15 +49,15 @@ export class ServerWsEndpoint extends AbstractWsEndpoint<ServerWsConnection> {
             server: this.httpServer,
             maxPayload: 1024 * 1024
         }).on('error', (err: Error) => {
-            this.logger.error('web socket server (wss) emitted error: %s', err)
+            logger.error('Encountered error (emitted by WebSocket.Server)', { err })
         }).on('listening', () => {
-            this.logger.trace('listening on %s', this.getUrl())
+            logger.trace('Started', { url: this.getUrl() })
         }).on('connection', (ws: WebSocket, request: http.IncomingMessage) => {
             const handshakeUUID = v4()
 
             this.handshakeTimeoutRefs[handshakeUUID] = setTimeout(() => {
                 ws.close(DisconnectionCode.FAILED_HANDSHAKE, `Handshake not received from connection behind UUID ${handshakeUUID}`)
-                this.logger.warn(`Handshake not received from connection behind UUID ${handshakeUUID}`)
+                logger.warn('Timed out waiting for handshake from connection', { handshakeUUID })
                 ws.terminate()
                 delete this.handshakeTimeoutRefs[handshakeUUID]
             }, this.handshakeTimer)
@@ -80,18 +83,20 @@ export class ServerWsEndpoint extends AbstractWsEndpoint<ServerWsConnection> {
                         } else {
                             const failedMessage = `Connection for node: ${peerId} has already been established, rejecting duplicate`
                             ws.close(DisconnectionCode.DUPLICATE_SOCKET, failedMessage)
-                            this.logger.warn(failedMessage)
+                            logger.warn('Reject duplicate connection (connection to peer has already been established)', {
+                                peerId
+                            })
                         }
                     } else {
-                        this.logger.trace('Expected a handshake message got: ' + data.toString())
+                        logger.trace('Received unexpected message (expected handshake message)', { message: data.toString() })
                     }
                 } catch (err) {
-                    this.logger.trace(err)
+                    logger.trace('startWsServer', { err })
                 }
             })
 
             ws.on('error', (err) => {
-                this.logger.warn('socket for "%s" emitted error: %s', otherNodeIdForLogging, err)
+                logger.warn('Encountered error (emitted by socket)', { otherNodeIdForLogging, err })
             })
         })
     }
@@ -118,7 +123,7 @@ export class ServerWsEndpoint extends AbstractWsEndpoint<ServerWsConnection> {
         })
 
         duplexStream.on('error', (error) => {
-            this.logger.error('Duplex stream error: ' + error.stack as string)
+            logger.error( 'Encountered error (emitted by DuplexStream)', { stack: error.stack as string })
         })
 
         ws.on('pong', () => {
@@ -150,11 +155,11 @@ export class ServerWsEndpoint extends AbstractWsEndpoint<ServerWsConnection> {
             }
             this.wss.close((err?) => {
                 if (err) {
-                    this.logger.error('error on closing websocket server: %s', err)
+                    logger.error('Encountered error (while closing WebSocket.Server)', { err })
                 }
                 this.httpServer.close((err?) => {
                     if (err) {
-                        this.logger.error('error closing http server: %s', err)
+                        logger.error('Encountered error (while closing httpServer)', { err })
                         reject(err)
                     } else {
                         resolve()
@@ -179,8 +184,7 @@ function cleanSocket(httpServer: http.Server | https.Server, config: UnixSocket)
         // rethrow if unexpected error
         if (!err.message.includes('EADDRINUSE')) { throw err }
 
-        staticLogger.info('socket in use, trying to recover: %s', config)
-        staticLogger.trace('checking if socket in use by another server')
+        logger.info('Try to recover used socket', { config })
         const clientSocket = new net.Socket()
         // socket will automatically close on error
         clientSocket.on('error', (err: any) => {
@@ -191,7 +195,7 @@ function cleanSocket(httpServer: http.Server | https.Server, config: UnixSocket)
 
             // No other server listening
             try {
-                staticLogger.trace('cleaning unused socket: %s', config)
+                logger.trace('Clean unused socket', { config })
                 fs.unlinkSync(config)
             } catch (unlinkErr) {
                 // ignore error if somehow file was already removed
@@ -206,7 +210,7 @@ function cleanSocket(httpServer: http.Server | https.Server, config: UnixSocket)
 
         clientSocket.once('connect', () => {
             // bad news if we are able to connect
-            staticLogger.error('Another server already running on socket: %s', config)
+            logger.error('Encountered unexpected reserved socket (another server already running?)', { config })
             process.exit(1)
         })
         clientSocket.connect({ path: config })
@@ -237,15 +241,15 @@ export async function startHttpServer(
     try {
         httpServer.listen(config)
         await once(httpServer, 'listening')
-        staticLogger.info(`listening on %s`, JSON.stringify(config))
+        logger.info('Listen', { details: JSON.stringify(config) })
     } catch (err) {
         // Kill process if started on host/port, else wait for Unix Socket to be cleaned up
         if (typeof config !== "string") {
-            staticLogger.error(err)
+            logger.error('Failed to start httpServer', err)
             process.exit(1)
         } else {
             await once(httpServer, 'listening')
-            staticLogger.info(`listening on %s`, JSON.stringify(config))
+            logger.info('Listen', { details: JSON.stringify(config) })
         }
     }
     return httpServer