+++ packages/broker/CHANGELOG.md
@@ -9,13 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add beneficiary address prompt to config wizard
+
 ### Changed
+
+- Info plugin provides more detailed diagnostic info
+- Websocket plugin includes more logging
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Websocket plugin cleans up subscriptions on client disconnect
 
 ### Security